refactor to map Elements instead of shape/txt
