+++ frontend/components/editor/Design.jsx
@@ -1,15 +1,8 @@
 import React from 'react';
-import Shape from './elements/Shape';
+import Element from './elements/Element';
 import styles from './Design.module.css';
 
 // const Shape = () => <div>I am a shape</div>;
-
-const Text = () => <div>I am a text</div>;
-
-const components = {
-  Shape,
-  Text,
-};
 
 // eslint-disable-next-line react/prefer-stateless-function
 class Design extends React.Component {
@@ -19,11 +12,7 @@ class Design extends React.Component {
     // debugger;
     return (
       <div className={styles.design} style={{ width: design.width, height: design.height }}>
-        {elements.map((element, index) => React.createElement(
-          components[element.elementableType],
-          { key: index, elementAttr: element.elementableAttributes, element },
-          null,
-        ))}
+        {elements.map((element) => <Element key={element.id} element={element} />)}
       </div>
     );
   }
